requirements.txt
flask==3.0.0
anthropic==0.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
pillow-heif>=0.16.0
python-docx==1.1.0
werkzeug>=3.0.0
